requirements.txt
azure-functions
tweepy>=4.10.0
azure-storage-blob